Revert "Change the `dataType` of the `timestampTimeoutAttribute` from IntegerType to LongType"
This reverts commit 57487e58116577fb09cc1255ca14100f94f901ab.

+++ sql/core/src/main/scala/org/apache/spark/sql/execution/streaming/state/FlatMapGroupsWithStateExecHelper.scala
@@ -152,7 +152,7 @@ object FlatMapGroupsWithStateExecHelper {
       shouldStoreTimestamp: Boolean) extends StateManagerImplBase(shouldStoreTimestamp) {
 
     private val timestampTimeoutAttribute =
-      AttributeReference("timeoutTimestamp", dataType = LongType, nullable = false)()
+      AttributeReference("timeoutTimestamp", dataType = IntegerType, nullable = false)()
 
     private val stateAttributes: Seq[Attribute] = {
       val encSchemaAttribs = toAttributes(stateEncoder.schema)
